+++ src/source/RazorEngine.Core/Templating/RazorEngineService.cs
@@ -122,23 +122,15 @@ namespace RazorEngine.Templating
         /// <summary>
         /// Releases managed resources used by this instance.
         /// </summary>
-        public void Dispose()
-        {
-            Dispose(true);
-            GC.SuppressFinalize(this);
-        }
-
-        /// <summary>
-        /// Releases managed resources used by this instance.
-        /// </summary>
         /// <param name="disposing">Are we explicitly disposing of this instance?</param>
-        protected virtual void Dispose(bool disposing)
+        protected override void Dispose(bool disposing)
         {
             if (!disposed && disposing)
             {
                 _config.CachingProvider.Dispose();
                 disposed = true;
             }
+            base.Dispose(disposing);
         }
 
         /// <summary>
@@ -259,6 +251,7 @@ namespace RazorEngine.Templating
         /// <param name="key"></param>
         /// <param name="modelType"></param>
         /// <param name="model"></param>
+        /// <param name="viewbag"></param>
         /// <returns></returns>
         internal ITemplate GetTemplate(ITemplateKey key, Type modelType, object model, DynamicViewBag viewbag)
         {