+++ src/test/Test.RazorEngine.Core/TestTypes/BaseTypes/NonGenericTemplateBase.cs
@@ -2,8 +2,14 @@
 
 namespace RazorEngine.Tests.TestTypes.BaseTypes
 {
+    /// <summary>
+    /// Test class.
+    /// </summary>
     public abstract class NonGenericTemplateBase : TemplateBase
     {
+        /// <summary>
+        /// Test class.
+        /// </summary>
         public string GetHelloWorldText()
         {
             return "Hello World";