+++ src/test/TestHelper/TestClass.cs
@@ -11,6 +11,9 @@ namespace TestHelper
     /// </summary>
     public class TestClass
     {
+        /// <summary>
+        /// Test property.
+        /// </summary>
         public string TestProperty
         {
             get { return "TestPropert"; }